WIP Darn react component won't work
