+++ frog/imports/ui/StudentView.jsx
@@ -9,6 +9,7 @@ import { objectize } from '../../lib/utils';
 import { createLogger } from '../api/logs';
 import { Sessions } from '../api/sessions';
 import { Activities } from '../api/activities';
+import { ActivityData, reactiveData } from '../api/activity_data';
 
 import { activity_types_obj } from '../activity_types';
 
@@ -18,7 +19,8 @@ const setStudentSession = (session_id) => {
 
 const SessionList = ( { sessions } ) => { return(
   <ul> { 
-    sessions.filter((session) => session.state=='CREATED').map((session) => 
+    //sessions.filter((session) => session.state=='CREATED').map((session) => 
+    sessions.map((session) => 
       <li key={session._id}>
         <button onClick={ () => setStudentSession(session._id) }>Join</button>
         {session.state}
@@ -28,17 +30,41 @@ const SessionList = ( { sessions } ) => { return(
   } </ul>
 )}
 
-const ActivityBody = ( { activity } ) => {
-  const Runner = ( { activity } ) => {
-    const activity_type = activity_types_obj[activity.activity_type]
+// should be separated into its own file
+const Runner = ( { activity } ) => {
+  const activity_type = activity_types_obj[activity.activity_type]
 
-    const logger = createLogger({
-      activity: activity._id, 
-      activity_type: activity.activity_type, 
-      user: Meteor.userId()
-    })
-    return(<activity_type.ActivityRunner config={activity.data} logger={logger}/>)
+  const logger = createLogger({
+    activity: activity._id, 
+    activity_type: activity.activity_type, 
+    user: Meteor.userId()
+  })
+
+  if(activity_type.meta.mode == 'collab') { 
+    const reactiveFn = reactiveData(1, 1, 1)
+    const runner = <activity_type.ActivityRunner 
+      config={activity.data} 
+      logger={logger} 
+      reactiveFn = {reactiveFn} /> 
+
+      // //        reactive: ActivityData.find({session_id: 1, activity_id: 1, group_id: 1}).fetch() 
+      // const ex = createContainer(() => {
+      //   return {
+      //     sessions: Sessions.find().fetch(),
+      //     user: Meteor.users.findOne({_id:Meteor.userId()})
+      //   }
+      // }, runner)
+    // runwrapped = ex
+  return runner
+
+  } else {
+    return <activity_type.ActivityRunner 
+      config={activity.data} 
+      logger={logger} />
   }
+}
+
+const ActivityBody = ( { activity } ) => {
   return(activity ? <Runner activity={activity}/>:<p>No activity selected</p>)
 }
