+++ op/op-random/src/index.js
@@ -1,21 +1,9 @@
-const splitAt = function(i, xs) {
-  var a = xs.slice(0, i);
-  var b = xs.slice(i, xs.length);
-  return [a, b];
-};
+import { shuffleList, splitAt, zipList } from 'frog-utils'
 
-const shuffle = function(xs) {
-  return xs.slice(0).sort(function() {
-    return .5 - Math.random();
-  });
-};
-
-const zip = function(xs) {
-  return xs[0].map(function(_,i) {
-    return xs.map(function(x) {
-      return x[i];
-    });
-  });
+export const meta = {
+  name: 'Random Operator',
+  type: 'social',
+  mode: 'one-pass'
 }
 
 export const config = {
@@ -30,6 +18,6 @@ export const config = {
 }
 
 // Obviously assumes even array
-export const operator = (names) => zip(splitAt(names.length/2, shuffle(names)))
+export const operator = (names) => zipList(splitAt(names.length/2, shuffleList(names)))
 
-export default { id: 'op-random', operator: operator, config: config }
+export default { id: 'op-random', operator: operator, config: config, meta: meta }
